requirements.txt
backoff==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
googleapis-common-protos==1.55.0
grpcio==1.44.0
idna==3.3
opentelemetry-api==1.9.1
opentelemetry-exporter-otlp-proto-grpc==1.9.1
opentelemetry-instrumentation==0.28b1
opentelemetry-proto==1.9.1
opentelemetry-sdk==1.9.1
opentelemetry-semantic-conventions==0.28b1
protobuf==3.19.4
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
